Consolidate ALTER statements and remove subquery within DELETE.
MySQL won't allow a DELETE with a subquery on the same table in case of
cycles, so we need to execute the subquery separately. We also combine
ALTER statements on the same table to let the database optimize the
operation.

+++ migrations/versions/084_mutable_drafts.py
@@ -11,7 +11,6 @@ revision = '10db12da2005'
 down_revision = '10a1129fe685'
 
 from alembic import op
-from sqlalchemy.dialects import mysql
 from sqlalchemy.sql import text
 
 import sqlalchemy as sa
@@ -22,20 +21,29 @@ def upgrade():
     op.add_column('actionlog',
                   sa.Column('extra_args', JSON(), nullable=True))
 
-    op.add_column('message',
-                  sa.Column('version', mysql.BINARY(16), nullable=True))
-
-    op.drop_constraint('message_ibfk_3', 'message', type_='foreignkey')
-
     conn = op.get_bind()
+
     conn.execute(text("""
-        DELETE FROM message
-        WHERE message.is_created = 1 AND message.is_draft = 1
-              AND message.id IN (
-                SELECT message.parent_draft_id from message
-                WHERE message.is_created = 1 AND message.is_draft = 1
-                    AND message.parent_draft_id IS NOT NULL)
-        """))
+        ALTER TABLE message ADD COLUMN version BINARY(16),
+                            DROP FOREIGN KEY message_ibfk_3
+                      """))
+
+    parent_drafts_ids = [id_ for id_, in conn.execute(text(
+        """
+        SELECT message.parent_draft_id from message
+        WHERE message.is_created = 1
+            AND message.is_draft = 1
+            AND message.parent_draft_id IS NOT NULL
+        """))]
+    print parent_drafts_ids
+    if parent_drafts_ids:
+        # delete old parent drafts
+        conn.execute(text("""
+            DELETE FROM message
+            WHERE message.is_created = 1 AND message.is_draft = 1
+            AND message.id IN :parent_drafts_ids"""),
+                     parent_drafts_ids=parent_drafts_ids)
+
     conn.execute(text("""
         UPDATE message SET message.version = message.public_id
         WHERE message.is_created = 1 AND message.is_draft = 1
